LayoutTests/storage/indexeddb/setversion-blocks-open-expected.txt

CONSOLE MESSAGE: The setVersion() method is non-standard and will be removed. Use the "upgradeneeded" event instead.
Test that open is blocked if a VERSION_CHANGE transaction is running

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

dbname = "setversion-blocks-open.html"
ver = 1
versionChangeComplete = false
indexedDB.open(dbname)

h1OpenOnSuccess():
h1 = event.target.result
h1.setVersion(String(ver++))
indexedDB.open(dbname)

h1SetVersionOnSuccess():

transactionOnComplete():
versionChangeComplete = true

h2OpenOnSuccess():
h2 = event.target.result
PASS versionChangeComplete is true
PASS successfullyParsed is true

TEST COMPLETE
